refactor: move EosModel into fluid
Put fluid related stuff together. Caused by issue with
circular import since neqsim wrapper uses fluid classes.
fluid.py will be moved to common next.
