+++ src/libecalc/dto/types.py
@@ -21,13 +21,6 @@ class ChartControlMarginUnit(str, Enum):
     PERCENTAGE = "PERCENTAGE"
 
 
-class EoSModel(str, Enum):
-    SRK = "SRK"
-    PR = "PR"
-    GERG_SRK = "GERG_SRK"
-    GERG_PR = "GERG_PR"
-
-
 class InterpolationType(str, Enum):
     LEFT = "LEFT"
     RIGHT = "RIGHT"
