+++ src/libecalc/presentation/yaml/mappers/fluid_mapper.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 from libecalc.dto import FluidComposition, FluidModel
-from libecalc.dto.types import EoSModel
+from libecalc.dto.models.compressor.fluid import EoSModel
 from libecalc.presentation.yaml.resource import Resources
 from libecalc.presentation.yaml.yaml_keywords import EcalcYamlKeywords
 from libecalc.presentation.yaml.yaml_types.models import YamlFluidModel
